#25 work on toggling marked comments - Not working

+++ Rawdata.Data/Scripts/functions.sql
@@ -173,12 +173,20 @@ RETURNS BOOLEAN AS $$
 	DECLARE
 		_marked BOOLEAN;
     BEGIN
-        IF NOT EXISTS (SELECT * FROM marked_posts WHERE "user_id" = _user_id AND post_id = _post_id) THEN
-            INSERT INTO marked_posts VALUES(_user_id, _post_id, _note);
-			_marked := true;
+		 IF _user_id IS NOT NULL THEN
+            IF _user_id IN (SELECT "id" FROM users) THEN
+				IF NOT EXISTS (SELECT * FROM marked_posts WHERE "user_id" = _user_id AND post_id = _post_id) THEN
+					INSERT INTO marked_posts VALUES(_user_id, _post_id, _note);
+					_marked := true;
+				ELSE
+					DELETE FROM marked_posts WHERE "user_id" = _user_id AND post_id = _post_id;
+					_marked := false;
+				END IF;
+			 ELSE
+                RAISE EXCEPTION 'search_marked_comments: User does not exist.';
+            END IF;
         ELSE
-            DELETE FROM marked_posts WHERE "user_id" = _user_id AND post_id = _post_id;
-			_marked := false;
+            RAISE EXCEPTION 'search_marked_comments: User ID required.';
         END IF;
 
 		RETURN _marked;
@@ -186,19 +194,19 @@ RETURNS BOOLEAN AS $$
 $$ LANGUAGE plpgsql;
 
 CREATE FUNCTION toggle_marked_comment(_user_id INTEGER, _comment_id INTEGER, _note TEXT = NULL)
-RETURNS BOOLEAN AS $$
-	DECLARE
-		_marked BOOLEAN;
-    BEGIN
-        IF NOT EXISTS (SELECT * FROM marked_comments WHERE "user_id" = _user_id AND comment_id = _comment_id) THEN
-            INSERT INTO marked_comments VALUES(_user_id, _comment_id, _note);
-			_marked := true;
-        ELSE
-            DELETE FROM marked_comments WHERE "user_id" = _user_id AND comment_id = _comment_id;
-			_marked := false;
-        END IF;
+RETURNS SETOF marked_comments AS $$
+	BEGIN
+			IF _user_id IS NULL THEN RAISE EXCEPTION 'toggle_marked_comment: User ID required.';
+			ELSIF _user_id NOT IN (SELECT "id" FROM users) THEN RAISE EXCEPTION 'toggle_marked_comment: User does not exist.';
+			ELSIF _comment_id IS NULL THEN RAISE EXCEPTION 'toggle_marked_comment: Comment ID required.';
+			ELSIF _comment_id NOT IN (SELECT "id" FROM comments) THEN RAISE EXCEPTION 'toggle_marked_comment: Comment does not exist.';
+			ELSIF NOT EXISTS (SELECT * FROM marked_comments WHERE "user_id" = _user_id AND comment_id = _comment_id) THEN 
+				INSERT INTO marked_comments VALUES(_user_id, _comment_id, _note);
+			ELSE 
+				DELETE FROM marked_comments WHERE "user_id" = _user_id AND comment_id = _comment_id;
+			END IF;
 
-		RETURN _marked;
+			RETURN QUERY SELECT * FROM marked_comments WHERE "user_id" = _user_id AND comment_id = _comment_id;
     END
 $$ LANGUAGE plpgsql;
